revert the lib size changes because they broke everything
