+++ src/tmb.cpp
@@ -1,0 +1,7 @@
+#ifdef WITH_LIBTMB
+#undef WITH_LIBTMB
+#endif
+
+#define TMB_PRECOMPILE
+
+#include <TMB.hpp>
