+++ src/tmb_includes.h
@@ -1,0 +1,9 @@
+#ifndef TMB_INCLUDES_H
+#define TMB_INCLUDES_H
+
+/* TMB */
+#define WITH_LIBTMB
+
+#include <TMB.hpp>
+
+#endif
